% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzim.R
\name{print.summary.mzim}
\alias{print.summary.mzim}
\title{Print method for summary.mzim objects}
\usage{
\method{print}{summary.mzim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.mzim}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for summary.mzim objects
}
