% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzim.R
\name{summary.mzim}
\alias{summary.mzim}
\title{Summary for mzim objects}
\usage{
\method{summary}{mzim}(object, robust = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{mzim}.}

\item{robust}{Logical. If TRUE, uses sandwich standard errors.}

\item{...}{Additional arguments.}
}
\description{
Summary for mzim objects
}
